Rename package, update documentation
